* SharedOverThereToo version 0.1.1
SharedOverThereToo allows a user to upload a file to multiple locations.
The criteria for creating folders in the content directory is the
filesize instead of latitude and longitude. The server-side PHP files
were modified to accomodate this change. This has not been tested on an
actual Android phone, as all the hosting was done locally. All of the
client-side files reflect this. The server-side files should be okay
though.

I recommend you back up your SharedHere 0.1.0 client directory and use
that if you plan to run this on an actual Android phone. I chose to
commit my client directory so that others could run it in exactly the
same environment I did.

A few notes:

1.) The .java files are configured to run on an Android 4.1 emulator. I
think I made it pretty clear what you would have to do in order to
remove the hard coded location values. None of the modifications I did
to the java files affect the App's functionality aside from this. In
short, you may not need them at all.

2.) There are still a few System.out.println messages there that I used
for debugging purposes. Most of them are commented out, but they don't
affect the performance of the app on the emulator, so I left them there.

3.) I ran a localhost MySQL server for testing purposes, hence the IP
address in place of where the SharedHere webservice URL was in
strings.xml. My SQL dump is included.